update mapfuncs and fix unit tests
