+++ controller/hybridgateway/watch/mapfuncs_httproute.go
@@ -4,14 +4,11 @@ import (
 	"context"
 
 	"github.com/go-logr/logr"
-	corev1 "k8s.io/api/core/v1"
-	discoveryv1 "k8s.io/api/discovery/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	configurationv1 "github.com/kong/kong-operator/v2/api/configuration/v1"
-	configurationv1alpha1 "github.com/kong/kong-operator/v2/api/configuration/v1alpha1"
 	"github.com/kong/kong-operator/v2/controller/hybridgateway/metadata"
 	gwtypes "github.com/kong/kong-operator/v2/internal/types"
 	"github.com/kong/kong-operator/v2/internal/utils/index"
@@ -62,106 +59,6 @@ func listHTTPRoutesForService(ctx context.Context, cl client.Client, svcNamespac
 		}
 	}
 	return requests, nil
-}
-
-// MapHTTPRouteForGateway returns a handler.MapFunc that, given a Gateway object,
-// lists all HTTPRoutes referencing that Gateway (via ParentRefs) using the GatewayOnHTTPRouteIndex.
-// It returns a slice of reconcile.Requests for each matching HTTPRoute, enabling efficient event handling
-// and reconciliation when a Gateway changes.
-func MapHTTPRouteForGateway(cl client.Client) handler.MapFunc {
-	return func(ctx context.Context, obj client.Object) []reconcile.Request {
-		gateway, ok := obj.(*gwtypes.Gateway)
-		if !ok {
-			return nil
-		}
-		requests, err := listHTTPRoutesForGateway(ctx, cl, gateway.Namespace, gateway.Name)
-		if err != nil {
-			return nil
-		}
-		return requests
-	}
-}
-
-// MapHTTPRouteForGatewayClass returns a handler.MapFunc that, given a GatewayClass object,
-// lists all Gateways referencing that GatewayClass (using GatewayClassOnGatewayIndex),
-// then for each Gateway, lists all HTTPRoutes referencing it (via ParentRefs and GatewayOnHTTPRouteIndex).
-// It returns a slice of reconcile.Requests for each matching HTTPRoute, enabling efficient event handling
-// and reconciliation when a GatewayClass changes.
-func MapHTTPRouteForGatewayClass(cl client.Client) handler.MapFunc {
-	return func(ctx context.Context, obj client.Object) []reconcile.Request {
-		gc, ok := obj.(*gwtypes.GatewayClass)
-		if !ok {
-			return nil
-		}
-
-		// List all Gateways that reference this GatewayClass using the index.
-		gateways := &gwtypes.GatewayList{}
-		err := cl.List(ctx, gateways, client.MatchingFields{
-			index.GatewayClassOnGatewayIndex: gc.Name,
-		})
-		if err != nil {
-			return nil
-		}
-
-		var requests []reconcile.Request
-		for _, gateway := range gateways.Items {
-			gwRequests, err := listHTTPRoutesForGateway(ctx, cl, gateway.Namespace, gateway.Name)
-			if err != nil {
-				return nil
-			}
-			requests = append(requests, gwRequests...)
-		}
-		return requests
-	}
-}
-
-// MapHTTPRouteForService returns a handler.MapFunc that, given a Service object,
-// lists all HTTPRoutes referencing that Service using the BackendServicesOnHTTPRouteIndex.
-// It returns a slice of reconcile.Requests for each matching HTTPRoute, enabling efficient event handling
-// and reconciliation when a Service changes.
-func MapHTTPRouteForService(cl client.Client) handler.MapFunc {
-	return func(ctx context.Context, obj client.Object) []reconcile.Request {
-		svc, ok := obj.(*corev1.Service)
-		if !ok {
-			return nil
-		}
-
-		requests, err := listHTTPRoutesForService(ctx, cl, svc.Namespace, svc.Name)
-		if err != nil {
-			return nil
-		}
-		return requests
-	}
-}
-
-// MapHTTPRouteForEndpointSlice returns a handler.MapFunc that, given an EndpointSlice object,
-// retrieves the owning Service and lists all HTTPRoutes referencing that Service using the BackendServicesOnHTTPRouteIndex.
-// It returns a slice of reconcile.Requests for each matching HTTPRoute, enabling efficient event handling
-// and reconciliation when an EndpointSlice changes.
-func MapHTTPRouteForEndpointSlice(cl client.Client) handler.MapFunc {
-	return func(ctx context.Context, obj client.Object) []reconcile.Request {
-		epSlice, ok := obj.(*discoveryv1.EndpointSlice)
-		if !ok {
-			return nil
-		}
-
-		// Get the Service that owns this EndpointSlice.
-		svcName, ok := epSlice.Labels[discoveryv1.LabelServiceName]
-		if !ok {
-			return nil
-		}
-		svc := &corev1.Service{}
-		err := cl.Get(ctx, client.ObjectKey{Namespace: epSlice.Namespace, Name: svcName}, svc)
-		if err != nil {
-			return nil
-		}
-
-		requests, err := listHTTPRoutesForService(ctx, cl, svc.Namespace, svc.Name)
-		if err != nil {
-			return nil
-		}
-		return requests
-	}
 }
 
 // MapHTTPRouteForReferenceGrant returns a handler.MapFunc that, given a ReferenceGrant object,
@@ -218,17 +115,6 @@ func MapHTTPRouteForReferenceGrant(cl client.Client) handler.MapFunc {
 		}
 		return requests
 	}
-}
-
-// kongResource is a type constraint that encompasses all Kong resource types
-// that can be mapped back to HTTPRoutes via annotations.
-type kongResource interface {
-	*configurationv1alpha1.KongUpstream |
-		*configurationv1alpha1.KongTarget |
-		*configurationv1alpha1.KongService |
-		*configurationv1alpha1.KongRoute |
-		*configurationv1.KongPlugin |
-		*configurationv1alpha1.KongPluginBinding
 }
 
 // MapHTTPRouteForKongResource returns a handler.MapFunc that, given a Kong resource object of type T,
@@ -292,7 +178,7 @@ func MapHTTPRouteForKongPlugin(cl client.Client) handler.MapFunc {
 		}
 
 		// Add requests for Plugins referencing the HTTPRoute via annotation.
-		requests := MapHTTPRouteForKongResource[*configurationv1.KongPlugin](cl)(ctx, obj)
+		requests := MapRouteForKongResource[*configurationv1.KongPlugin](cl)(ctx, obj)
 		return append(requests, indexRequests...)
 	}
 }
